Retake Exam Aug 2022

+++ Java OOP/JavaOOPExamPractise/RetakeExamAug2022/src/goldDigger/models/discoverer/Anthropologist.java
@@ -2,7 +2,7 @@ package goldDigger.models.discoverer;
 
 public class Anthropologist extends BaseDiscoverer {
 
-    private static final int INITIAL_ENERGY_UNITS = 40;
+    private static final double INITIAL_ENERGY_UNITS = 40;
 
     public Anthropologist(String name) {
         super(name, INITIAL_ENERGY_UNITS);
