+++ Java OOP/JavaOOPExamPractise/RetakeExamAug2022/src/goldDigger/models/discoverer/Archaeologist.java
@@ -2,7 +2,7 @@ package goldDigger.models.discoverer;
 
 public class Archaeologist extends BaseDiscoverer {
 
-    private static final int INITIAL_ENERGY_UNITS = 60;
+    private static final double INITIAL_ENERGY_UNITS = 60;
 
     public Archaeologist(String name) {
         super(name, INITIAL_ENERGY_UNITS);
